Renamed Kekiri.Examples to Kekiri.Examples.NUnit
