+++ src/Kekiri.Examples.NUnit/Calculator/Division/Divide_by_zero.cs
