+++ src/Kekiri.Examples.NUnit/Calculator/Subtraction/Subtracting_two_numbers.cs
