+++ src/Kekiri.Examples.NUnit/Kekiri.Examples.NUnit.csproj
@@ -3,8 +3,9 @@
   <PropertyGroup>
     <TargetFramework>netcoreapp1.1</TargetFramework>
     <OutputType>Exe</OutputType>
-    <AssemblyName>Kekiri.Examples</AssemblyName>
-    <IsPackage>false</IsPackage>
+    <AssemblyName>Kekiri.Examples.NUnit</AssemblyName>
+    <RootNamespace>$(AssemblyName)</RootNamespace>
+    <IsPackable>false</IsPackable>
   </PropertyGroup>
 
   <ItemGroup>
