+++ src/Kekiri.Examples.NUnit/Typed_context.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Kekiri.TestRunner.NUnit;
 using NUnit.Framework;
 
-namespace Kekiri.Examples
+namespace Kekiri.Examples.NUnit
 {
     public class Typed_context : Scenario<CalculatorContext>
     {
