+++ src/Kekiri.Examples.Xunit/Kekiri.Examples.Xunit.csproj
@@ -1,20 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <TargetFramework>netcoreapp1.1</TargetFramework>
     <DebugType>portable</DebugType>
     <AssemblyName>Kekiri.Examples.Xunit</AssemblyName>
-    <IsPackage>false</IsPackage>
+    <RootNamespace>$(AssemblyName)</RootNamespace>
+    <IsPackable>false</IsPackable>
   </PropertyGroup>
-
   <ItemGroup>
     <ProjectReference Include="../Kekiri.TestRunner.Xunit/Kekiri.TestRunner.Xunit.csproj" />
   </ItemGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.*" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.2.*" />
     <PackageReference Include="xunit" Version="2.2.*" />
   </ItemGroup>
-
 </Project>
