+++ src/Kekiri.TestRunner.NUnit/Kekiri.TestRunner.NUnit.csproj
@@ -6,6 +6,7 @@
     <TargetFramework>netstandard1.6</TargetFramework>
     <DebugType>portable</DebugType>
     <AssemblyName>Kekiri.TestRunner.NUnit</AssemblyName>
+    <RootNamespace>$(AssemblyName)</RootNamespace>
     <PackageId>Kekiri.TestRunner.NUnit</PackageId>
     <PackageTags>bdd;scenario;test;gherkin;cucumber;pickles;nunit;specflow</PackageTags>
     <PackageReleaseNotes>Test runners no longer require inheritance</PackageReleaseNotes>
