+++ src/Kekiri/ScenarioBase.cs
@@ -351,7 +351,6 @@ namespace Kekiri
 
         internal virtual IReportTarget CreateReportTarget()
         {
-            return DebugReportTarget.GetInstance();
             return CompositeReportTarget.GetInstance();
         }
     }
